requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pymongo==3.12.0
redis
python-dotenv
pytest